// src/common/exceptions/error-codes.enum.ts
export enum ErrorCode {
  VALIDATION_ERROR = 'VALIDATION_ERROR',
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
  RECURSO_NO_ENCONTRADO = 'RECURSO_NO_ENCONTRADO',
  PERMISO_DENEGADO = 'PERMISO_DENEGADO',
  CORREO_YA_REGISTRADO = 'CORREO_YA_REGISTRADO',
  PAIS_REQUERIDO_OPERADOR = 'PAIS_REQUERIDO_OPERADOR',
  PAIS_NO_ENCONTRADO = 'PAIS_NO_ENCONTRADO',
  TOKEN_ACTIVACION_INVALIDO = 'TOKEN_ACTIVACION_INVALIDO',
  TOKEN_ACTIVACION_EXPIRADO = 'TOKEN_ACTIVACION_EXPIRADO',
  TOKEN_ACTIVACION_USADO = 'TOKEN_ACTIVACION_USADO',
  USUARIO_YA_ACTIVO = 'USUARIO_YA_ACTIVO',
  AUTH_CREDENCIALES_INVALIDAS = 'AUTH_CREDENCIALES_INVALIDAS',
  AUTH_CUENTA_BLOQUEADA = 'AUTH_CUENTA_BLOQUEADA',
  AUTH_PRE_AUTH_INVALIDO = 'AUTH_PRE_AUTH_INVALIDO',
  AUTH_SESION_INVALIDA = 'AUTH_SESION_INVALIDA',
  PAIS_SESION_INVALIDO = 'PAIS_SESION_INVALIDO',
  PAIS_SESION_REQUERIDO = 'PAIS_SESION_REQUERIDO',
  AUTH_TOKEN_EXPIRADO = 'AUTH_TOKEN_EXPIRADO',
  USUARIO_NO_ENCONTRADO = 'USUARIO_NO_ENCONTRADO',
  USUARIO_NO_ACTIVO = 'USUARIO_NO_ACTIVO',
  CUENTA_NO_BLOQUEADA = 'CUENTA_NO_BLOQUEADA',
  PAIS_DUPLICADO = 'PAIS_DUPLICADO',
  UBICACION_NO_ENCONTRADA = 'UBICACION_NO_ENCONTRADA',
  UBICACION_DUPLICADA = 'UBICACION_DUPLICADA',
  CONFIGURACION_NO_ENCONTRADA = 'CONFIGURACION_NO_ENCONTRADA',
  CONFIGURACION_VALOR_INVALIDO = 'CONFIGURACION_VALOR_INVALIDO',
  CLIENTE_NO_ENCONTRADO = 'CLIENTE_NO_ENCONTRADO',
  CONTACTO_NO_ENCONTRADO = 'CONTACTO_NO_ENCONTRADO',
  CONTACTO_GENERICO_NO_ELIMINABLE = 'CONTACTO_GENERICO_NO_ELIMINABLE',
  CONTACTO_GENERICO_NO_EDITABLE = 'CONTACTO_GENERICO_NO_EDITABLE',
  RELACIONAMIENTO_NO_ENCONTRADO = 'RELACIONAMIENTO_NO_ENCONTRADO',
  RELACIONAMIENTO_FECHA_REUNION_INVALIDA = 'RELACIONAMIENTO_FECHA_REUNION_INVALIDA',
  SEGMENTO_OTRO_REQUERIDO = 'SEGMENTO_OTRO_REQUERIDO',
  CARGA_MASIVA_DESHABILITADA = 'CARGA_MASIVA_DESHABILITADA',
  CARGA_MASIVA_FORMATO_INVALIDO = 'CARGA_MASIVA_FORMATO_INVALIDO',
  PARAMETRO_NO_ENCONTRADO = 'PARAMETRO_NO_ENCONTRADO',
  PARAMETRO_DUPLICADO = 'PARAMETRO_DUPLICADO',
  PROCESO_NO_ENCONTRADO = 'PROCESO_NO_ENCONTRADO',
  PROCESO_ESTADO_INVALIDO = 'PROCESO_ESTADO_INVALIDO',
  PROCESO_INDICADORES_INCOMPLETOS = 'PROCESO_INDICADORES_INCOMPLETOS',
  PROCESO_INDICADORES_VACIOS_SIN_CONFIRMACION = 'PROCESO_INDICADORES_VACIOS_SIN_CONFIRMACION',
  PROCESO_FECHAS_SIN_INDICADORES = 'PROCESO_FECHAS_SIN_INDICADORES',
  PROCESO_EMPRESA_INVALIDA = 'PROCESO_EMPRESA_INVALIDA',
  PROCESO_RFI_FECHA_INVALIDA = 'PROCESO_RFI_FECHA_INVALIDA',
  TAREA_NO_ENCONTRADA = 'TAREA_NO_ENCONTRADA',
  TAREA_NO_APLICA = 'TAREA_NO_APLICA',
  TAREA_EVIDENCIA_REQUERIDA = 'TAREA_EVIDENCIA_REQUERIDA',
  VALIDACION_NO_ENCONTRADA = 'VALIDACION_NO_ENCONTRADA',
  VALIDACION_COMENTARIO_REQUERIDO = 'VALIDACION_COMENTARIO_REQUERIDO',
  VALIDACION_AVANCE_INCOMPLETO = 'VALIDACION_AVANCE_INCOMPLETO',
  SOLICITUD_ELIMINACION_NO_ENCONTRADA = 'SOLICITUD_ELIMINACION_NO_ENCONTRADA',
  SOLICITUD_ELIMINACION_YA_RESUELTA = 'SOLICITUD_ELIMINACION_YA_RESUELTA',
  PROYECCION_NO_ENCONTRADA = 'PROYECCION_NO_ENCONTRADA',
  PROYECCION_ORIGEN_DUPLICADA = 'PROYECCION_ORIGEN_DUPLICADA',
  PROYECCION_RESULTANTE_DUPLICADA = 'PROYECCION_RESULTANTE_DUPLICADA',
  NOTIFICACION_NO_ENCONTRADA = 'NOTIFICACION_NO_ENCONTRADA',
}